+++ geocat/comp/__init__.py
@@ -33,6 +33,8 @@ from .meteorology import (
     saturation_vapor_pressure_slope,
     delta_pressure,
     dpres_plev,
+    zonal_meridional_psi,
+    zonal_mpsi,
 )
 from .spherical import decomposition, recomposition, scale_voronoi
 from .stats import eofunc, eofunc_eofs, eofunc_pcs, eofunc_ts, pearson_r, nmse