+++ docs/release-notes.rst
@@ -15,6 +15,17 @@ Internal Changes
 ^^^^^^^^^^^^^^^^
 * Add dask compatibility testing for all functionality by `Anissa Zacharias`_ in (:pr:`819`)
 
+
+vYYYY.MM.## (unreleased)
+------------------------
+This release has updates to our internal structure to ingest UXarray datasets in a new zonal mean meridional stream function for data on unstructured grids using UXarray in ``meteorology.zonal_meridional_psi``. There are also two new helper functions ``gc_util._find_coord`` and ``gc_util._find_optional_coord``
+
+New Features
+^^^^^^^^^^^^
+* Zonal mean meridional stream function for data on unstructured grids using
+UXarray in ``meteorology.zonal_meridional_psi`` by `Julia Kent`_ in (:pr:`773`)
+
+
 v2026.01.0 (January 27, 2026)
 -----------------------------
 This release updates ``calendar_average`` and ``climate_anomaly`` to allow for